[Chore] APIEndpoint, APIHeader, NetworkError 분리 (#91)
